Add `goToEditMove` action in recording mode hook
Also tested this function
Also Split apart the recording mode tests from the others
